move some classes to bl module, add methods
