+++ codexbl/src/main/java/so/codex/codexbl/interactors/ISignInInteractor.kt
@@ -1,5 +1,8 @@
 package so.codex.codexbl.interactors
 
+import io.reactivex.Single
+import so.codex.codexbl.entity.UserAuth
+
 interface ISignInInteractor {
-    
+    fun signIn(userAuth: UserAuth): Single<Boolean>
 }
