+++ codexbl/src/main/java/so/codex/codexbl/output/CoreApi.kt
@@ -1,0 +1,5 @@
+package so.codex.codexbl.output
+
+class CoreApi {
+
+}
