+++ codexbl/src/main/java/so/codex/codexbl/output/entity/RequestEntity.kt
@@ -1,3 +1,3 @@
-package so.codex.codexsource.base
+package so.codex.codexbl.output.entity
 
 data class RequestEntity(val query: String, val params: Any)
