+++ codexbl/src/main/java/so/codex/codexbl/output/interfaces/IAuthApi.kt
@@ -1,0 +1,9 @@
+package so.codex.codexbl.output.interfaces
+
+import io.reactivex.Single
+import so.codex.codexbl.output.entity.AuthEntity
+import so.codex.codexbl.output.response.TokenResponse
+
+interface IAuthApi {
+    fun login(auth: AuthEntity): Single<TokenResponse>
+}
