+++ codexbl/src/main/java/so/codex/codexbl/output/response/TokenResponse.kt
@@ -1,3 +1,3 @@
-package so.codex.codexsource.response
+package so.codex.codexbl.output.response
 
 class TokenResponse(val accessToken: String, val refreshToken: String)
