+++ codexsource/src/main/java/so/codex/codexsource/api/AuthApiMethods.kt
@@ -2,8 +2,8 @@ package so.codex.codexsource.api
 
 import io.reactivex.Single
 import retrofit2.http.GET
-import so.codex.codexsource.base.AuthEntity
-import so.codex.codexsource.response.TokenResponse
+import so.codex.codexbl.output.entity.AuthEntity
+import so.codex.codexbl.output.response.TokenResponse
 
 interface AuthApiMethods {
     @GET("/graphql")
