--- codexsource/src/main/java/so/codex/codexsource/base/AuthEntity.kt
@@ -1,3 +1,0 @@
-package so.codex.codexsource.base
-
-data class AuthEntity(val email: String, val password: String)
